fix: fix lint
