+++ packages/client/app/page.tsx
@@ -1,24 +1,5 @@
-import Image from 'next/image'
-import { SNSButton as SNSLogin } from '@gds/component'
-import BackGroundImage from '../public/asset/Background.svg'
-import * as styles from './signin.css'
+import { ClientSigninPage } from 'src/clientPages/signin/ui/ClientSigninPage'
 
 export default function Signin() {
-   return (
-      <div className={styles.Wrapper}>
-         <div className={styles.TextWrapper}>
-            <div className={styles.TitleTxt}>
-               <Image src={BackGroundImage} alt="background image" />
-            </div>
-         </div>
-         <div className={styles.LoginBtnWrapper}>
-            <div>
-               <SNSLogin socialLoginType="kakao">카카오로 시작하기</SNSLogin>
-            </div>
-            <div style={{ marginTop: '16px' }}>
-               <SNSLogin socialLoginType="naver">네이버로 시작하기</SNSLogin>
-            </div>
-         </div>
-      </div>
-   )
+   return <ClientSigninPage />
 }
