+++ packages/client/src/clientPages/signin/ui/index.ts
@@ -1,0 +1,1 @@
+export { ClientSignInPage } from './ClientSigninPage'
