+++ packages/client/src/design-system/component/SNSButton/SNSButton.tsx
@@ -1,19 +1,25 @@
-import React from 'react'
+import { ButtonHTMLAttributes, ReactNode } from 'react'
 import * as styles from './SNSButton.css'
 import { IconKakao } from '../../icon/default/IconKakao'
 import { IconNaver } from '../../icon/default/IconNaver'
 
-interface SNSButtonProps {
+interface SNSButtonProps extends ButtonHTMLAttributes<HTMLButtonElement> {
    socialLoginType: 'kakao' | 'naver'
-   children: React.ReactNode
+   children: ReactNode
 }
 
-export function SNSButton({ socialLoginType, children }: SNSButtonProps) {
+export function SNSButton({
+   socialLoginType,
+   children,
+   ...props
+}: SNSButtonProps) {
    return (
-      <div
+      <button
          className={styles.Wrapper({
             backgroundColor: socialLoginType,
          })}
+         type="button"
+         {...props}
       >
          <div className={styles.LogoBox}>
             {socialLoginType === 'kakao' ? <IconKakao /> : <IconNaver />}
@@ -25,6 +31,6 @@ export function SNSButton({ socialLoginType, children }: SNSButtonProps) {
          >
             <span>{children}</span>
          </div>
-      </div>
+      </button>
    )
 }
