+++ packages/gmi-design-system/src/component/SNSButton/SNSButton.tsx
@@ -3,12 +3,12 @@ import * as styles from './SNSButton.css'
 import { IconKakao } from '../../icon/default/IconKakao'
 import { IconNaver } from '../../icon/default/IconNaver'
 
-interface LoginProps {
+interface SNSButtonProps {
   socialLoginType: 'kakao' | 'naver'
   children: React.ReactNode
 }
 
-const SNSLogin = ({ socialLoginType, children }: LoginProps) => {
+export const SNSButton = ({ socialLoginType, children }: SNSButtonProps) => {
   return (
     <div
       className={styles.Wrapper({
@@ -28,5 +28,3 @@ const SNSLogin = ({ socialLoginType, children }: LoginProps) => {
     </div>
   )
 }
-
-export default SNSLogin
