--- packages/gmi-design-system/src/component/SignIn/SNSLogin.stories.tsx
@@ -1,18 +1,0 @@
-import SNSLogin from './SNSLogin'
-
-export default {
-  title: 'Component/LoginButton',
-}
-
-export function LoginButton() {
-  return (
-    <div style={{ width: '358px' }}>
-      <div>
-        <SNSLogin SNS="카카오" type="시작하기" />
-      </div>
-      <div style={{ marginTop: '16px' }}>
-        <SNSLogin SNS="네이버" type="시작하기" />
-      </div>
-    </div>
-  )
-}
