+++ packages/gmi-design-system/src/component/index.ts
@@ -11,6 +11,9 @@ export type { ButtonProps } from './Button/Button'
 export { Select } from './Select/Select'
 export type { SelectProps } from './Select/Select'
 
+export { SNSButton } from './SNSButton/SNSButton'
+export type { SNSButtonProps } from './SNSButton/SNSButton'
+
 export { IconButton } from './IconButton/IconButton'
 export type { IconButtonProps } from './IconButton/IconButton'
