+++ packages/gmi-design-system/src/icon/default/index.ts
@@ -13,12 +13,10 @@ export { IconCheckCircleFilled } from './IconCheckCircleFilled'
 export { IconCredit } from './IconCredit'
 export { IconEdit } from './IconEdit'
 export { IconHuman } from './IconHuman'
-export { IconLikesDefault } from './IconLikesDefault'
-export { IconLogo } from './IconLogo'
+export { IconSearch } from './IconSearch'
 export { IconKakao } from './IconKakao'
 export { IconNaver } from './IconNaver'
 export { IconLogoFilled } from './IconLogoFilled'
 export { IconMainClear } from './IconMainClear'
 export { IconPlus } from './IconPlus'
-export { IconSearch } from './IconSearch'
 export { IconThumbUpFilled } from './IconThumbUpFilled'
